Spark Funds_Investment_EDA
An in-depth analysis using statistics and exploratory data analysis concepts to detect global trends in investments and a deep dive into mass investor sentiments. Variate analysis of Investment types, countries, sectors and the key factors leading to such investments. The goal was to identify the best country, sector to invest in for a specific time in the future.